pyproject.toml
[tool.poetry]
name = "beam-suntory"
version = "1.0.7"
description = ""
authors = ["petro <xavrais312@gmail.com>", "andrii <andrey.berezyanskiy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
flask-migrate = "^4.0.4"
flask-wtf = "^1.1.1"
flask-mail = "^0.9.1"
flask-login = "^0.6.2"
python-dotenv = "^1.0.0"
click = "^8.1.3"
email-validator = "^1.3.1"
psycopg2-binary = "^2.9.5"
pydantic = "^2.3.0"
alchemical = "^0.7.1"
pandas = "^2.0.3"
pydantic-settings = "^2.0.3"
beautifulsoup4 = "^4.12.2"
filetype = "^1.2.0"
pillow = "^10.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
flake8 = "^4.0.1"
black = "^22.3.0"

[tool.poetry.group.dev.dependencies]
faker = "^18.3.1"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
